% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_bispectra.R
\name{cross_bicoherence}
\alias{cross_bicoherence}
\title{Estimate cross-coherence from time series data.}
\usage{
cross_bicoherence(x, y, z = y, dft_given = FALSE)
}
\arguments{
\item{x}{Given 1st time series, as a data frame or matrix with which columns
correspond to sampled stretches.}

\item{y}{Given 2nd time series, with the same dimension as x.}

\item{z}{Optional 3rd time series, with the same dimension as x (and thus as y).
If omitted, \code{y} is used instead.}

\item{dft_given}{If TRUE, suppose that DFTs is given instead of time series
data and skip the fast fourier transform. Default: FALSE.}
}
\value{
A data frame including the following columns:
\describe{
\item{f1:}{
The first elements of frequency pairs.
}
\item{f2:}{
The second elements of frequency pairs.
}
\item{value:}{
The estimated value of (magnitude-squared) cross-bicoherence at the respective
frequency pair.
}
}
}
\description{
Estimate cross-coherence from three real-valued time series data.
}
\references{
K. S. Lii and K. N. Helland. 1981. Cross-Bispectrum Computation and Variance Estimation. ACM Trans. Math. Softw. 7, 3 (September 1981), 284–294. DOI:https://doi.org/10.1145/355958.35596
}
