% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions.R
\name{.blackman_window}
\alias{.blackman_window}
\title{Blackman window function}
\usage{
.blackman_window(x)
}
\arguments{
\item{x}{A real number in [0, 1].}
}
\value{
A real number in [0, 1].
}
\description{
Calculate
\deqn{\frac{42}{100} - \frac{\cos(2 \pi x)}{2} + \frac{8 \cos(4 \pi x)}{100}.}{%
0.42 - 0.5 * cos(2 * pi * x) + 0.08 * cos(4 * pi * x).
}
}
\seealso{
\code{\link{.hamming_window}} and \code{\link{.hann_window}}.
}
\keyword{internal}
