% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions.R
\name{.hamming_window}
\alias{.hamming_window}
\title{Hamming window function}
\usage{
.hamming_window(x)
}
\arguments{
\item{x}{A real number in [0, 1].}
}
\value{
A real number in [0, 1].
}
\description{
Calculate
\deqn{\frac{25}{46} - \frac{21}{46} \cos(2 \pi x).}{%
25/46 - (21/46) * cos(2 * pi * x).
}
}
\seealso{
\code{\link{.hann_window}} and \code{\link{.blackman_window}}.
}
\keyword{internal}
