% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window_functions.R
\name{.hann_window}
\alias{.hann_window}
\title{Hann window function}
\usage{
.hann_window(x)
}
\arguments{
\item{x}{A real number in [0, 1].}
}
\value{
A real number in [0, 1].
}
\description{
Calculate
\deqn{\frac{1 - \cos(2 * \pi * x)}{2}.}{%
0.5 * (1 - cos(2 * pi * x)).
}
}
\seealso{
\code{\link{.hamming_window}} and \code{\link{.blackman_window}}.
}
\keyword{internal}
