% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_matching.R
\name{mode_matching}
\alias{mode_matching}
\title{Estimate cross-bicoherence's empirical null distribution by a mode matching method}
\usage{
mode_matching(xbc, t_max = NULL, d = 0.001)
}
\arguments{
\item{xbc}{cross-bicoherence, returned from \code{cross_bicoherence}.}

\item{t_max}{the upper limit of interval \deqn{S_0}, see the reference.}

\item{d}{the bin width of the tuning parameter.}
}
\description{
Estimate false discovery rate by fitting scaled chi-squared distribution as an empirical null of cross-bicoherence with Schwartzman's mode matching method.
}
\references{
Schwartzman, Armin. “Empirical Null and False Discovery Rate Inference for Exponential Families.” Annals of Applied Statistics 2, no. 4 (December 2008): 1332–59. https://doi.org/10.1214/08-AOAS184.
}
