% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{three_channel_model}
\alias{three_channel_model}
\title{A three-channel model of quadratic phase coupling}
\usage{
three_channel_model(
  f1,
  f2,
  f3,
  num_samples = 256,
  num_observations = 100,
  Fcoef1 = 1.2,
  Fcoef2 = 0.7,
  Fcoef3 = 0.8,
  Qcoef = 0.3,
  noise_sd = 1
)
}
\arguments{
\item{f1}{A function of period \eqn{2 \pi}{2 * pi} for the first channel.}

\item{f2}{A function of period \eqn{2 \pi}{2 * pi} for the second channel.}

\item{f3}{A function of period \eqn{2 \pi}{2 * pi} for the third channel.}

\item{num_samples}{The number of sampling points in an observation.}

\item{num_observations}{The number of observations.}

\item{Fcoef1}{The scaling factor of the period for a given periodic function
of the first channel.}

\item{Fcoef2}{The scaling factor of the period for a given periodic function
of the second channel.}

\item{Fcoef3}{The scaling factor of the period for a given periodic function
of the third channel.}

\item{Qcoef}{The coefficient applied to the multiplicative component of the
third channel.}

\item{noise_sd}{The standard deviation of a Gaussian noise perturbing
samples.}
}
\value{
A list of three data frames: \code{c1}, \code{c2}, and \code{c3}.
Each element has \code{num_observations} columns and \code{num_samples} rows.
}
\description{
Simulate observations by a three-channel model of quadratic phase coupling.
}
\details{
Given three periodic functions, this function generate a list of three data
frames in which each column represents a simulated observation at a channel.
The phase is chosen at random from \eqn{[0, 2 \pi]}{[0, 2 * pi]} for each
observation and each channel.
}
\examples{
sawtooth <- function(r) {
    x <- r/(2*pi)
    x - floor(x)
}
data <- three_channel_model(cos, sin, sawtooth)

}
